Remove Line Works for Second Channel
Fixes #30

Remove Line.vi works by first searching for the channel we are looking to delete by name and then removing that channel object if it exists. There was incorrect logic that skipped removing modules if the index of the channel found was 1. This was updated to skip removing any channels if the index is -1 (not found) which should be the intended behavior.