Fix64
Fixed #64

Changed from string.vi in the channel class to not scan from string and instead compare against a case structure and pass out the expected value.

The best way to do this would be to change the channel version but this functionality is not available at the moment and we would have to change the framework to handle that.